- 맨마지막 스크롤시 로딩바가 표시되면서 다음페이지를 가져오는데 반복호출이 가능한 버그수정 - MY디스크 삭제,복원시 체크박스와 스크롤에 오류생기며, 비정상작동되는 버그수정

+++ MAMBrowser/BLL/PrivateFileBLL.cs
@@ -121,6 +121,8 @@ namespace MAMBrowser.BLL
             foreach (var seq in seqList)
             {
                 var fileData = Get(seq);
+                if (fileData == null)
+                    continue;
                 if (File.Exists(fileData.FilePath))
                 {
                     _fileProtocol.Delete(fileData.FilePath);
